BUG-107210 - [ZEPPELIN-3539] Add jceks stored password support for LDAP
This is to add support for storing password in jceks for LDAP (realm).

If the hadoopSecurityCredentialPath path is present and not empty in the shiro.ini, then the password is read from the keystore file and it need not be stored inside the shiro.ini file.

[Improvement | Feature]

* [x] - Add documentation

* [ZEPPELIN-3539](https://issues.apache.org/jira/browse/ZEPPELIN-3539)

Create a keystore file using the hadoop credential command line
```
hadoop credential create ldapRealm.systemPassword -provider jceks://file/user/zeppelin/conf/zeppelin.jceks
```
Change the following values in the Shiro.ini file, and uncomment the line:
```
ldapRealm.hadoopSecurityCredentialPath = jceks://file/user/zeppelin/conf/zeppelin.jceks
```

* Does the licenses files need update? N/A
* Is there breaking changes for older versions? N/A
* Does this needs documentation? N/A

Author: Prabhjyot Singh <prabhjyotsingh@gmail.com>

Closes #3018 from prabhjyotsingh/ZEPPELIN-3539 and squashes the following commits:

4dcbaabb4 [Prabhjyot Singh] zeppelin server should not start if invaild configuration is applied
228022db3 [Prabhjyot Singh] doc for hadoopSecurityCredentialPath
1428b3d1b [Prabhjyot Singh] [ZEPPELIN-3539] Add jceks stored password support for LDAP

Change-Id: Id50532a0b05de1572efd7fb5dc8a67777ca66fb8